Factor out feature hover functionality.
