+++ src/components/ol/Map.vue
@@ -16,7 +16,6 @@ import Projection from 'ol/proj/Projection';
 import TileGrid from 'ol/tilegrid/TileGrid';
 import { register as olproj4 } from 'ol/proj/proj4';
 import { get as getProj } from 'ol/proj';
-import Overlay from 'ol/Overlay';
 import { GPX, GeoJSON, IGC, KML, TopoJSON } from 'ol/format';
 import { Vector as VectorLayer } from 'ol/layer';
 import { Vector as VectorSource } from 'ol/source';
@@ -26,6 +25,7 @@ import { WguEventBus } from '../../WguEventBus.js';
 import { LayerFactory } from '../../factory/Layer.js';
 import LayerUtil from '../../util/Layer';
 import PermalinkController from './PermalinkController';
+import HoverController from './HoverController';
 import MapInteractionUtil from '../../util/MapInteraction';
 import ViewAnimationUtil from '../../util/ViewAnimation';
 import { ColorTheme } from '../../mixins/ColorTheme'
@@ -82,7 +82,7 @@ export default {
       me.setOlButtonColor();
 
       // initialize map hover functionality
-      me.setupMapHover();
+      this.hoverController = new HoverController(me.map);
     }, 200);
   },
   destroyed () {
@@ -263,70 +263,7 @@ export default {
         });
       }
     },
-    /**
-     * Initializes the map hover functionality:
-     * Adds a little tooltip like DOM element, wrapped as OL Overlay to the
-     * map.
-     * Registers a 'pointermove' event on the map and shows the layer's
-     * 'hoverAttribute' if the layer is configured as 'hoverable'
-     */
-    setupMapHover () {
-      const me = this;
-      const map = me.map;
-      let overlayEl;
 
-      // create a span to show map tooltip
-      overlayEl = document.createElement('span');
-      overlayEl.classList.add('wgu-hover-tooltiptext');
-      overlayEl.classList.add('v-sheet');
-      overlayEl.classList.add(this.isDarkTheme ? 'theme--dark' : 'theme--light');
-      map.getTarget().append(overlayEl);
-
-      me.overlayEl = overlayEl;
-
-      // wrap the tooltip span in a OL overlay and add it to map
-      me.overlay = new Overlay({
-        element: overlayEl,
-        stopEvent: false,
-        className: 'wgu-hover-ol-overlay',
-        autoPan: true,
-        autoPanAnimation: {
-          duration: 250
-        }
-      });
-      map.addOverlay(me.overlay);
-
-      // show tooltip if a hoverable feature gets hit with the mouse
-      map.on('pointermove', me.onPointerMove, me);
-    },
-    /**
-     * Shows the hover tooltip on the map if an appropriate feature of a
-     * 'hoverable' layer was hit with the mouse.
-     *
-     * @param  {Object} event The OL event for pointermove
-     */
-    onPointerMove (event) {
-      const me = this;
-      const map = me.map;
-      const overlayEl = me.overlayEl;
-      let hoverAttr;
-      const features = map.getFeaturesAtPixel(event.pixel, { layerFilter: (layer) => {
-        if (layer.get('hoverable')) {
-          hoverAttr = layer.get('hoverAttribute');
-        }
-        return layer.get('hoverable');
-      } });
-      if (!features || features.length === 0 || !hoverAttr) {
-        hoverAttr = null;
-        overlayEl.innerHTML = null;
-        me.overlay.setPosition(undefined);
-        return;
-      }
-      const feature = features[0];
-      var attr = feature.get(hoverAttr);
-      overlayEl.innerHTML = attr;
-      me.overlay.setPosition(event.coordinate);
-    },
     /**
      * Initializes the geodata drag-drop functionality:
      * Adds the ol/interaction/DragAndDrop to the map and draws the dropped
